v3.1.3: scope narrowing, Gemma4 per-layer KV, BF16 kernels, Qwen3.5 MoE decode
Architecture:
- Remove LLaMA architecture (unsupported since scope narrowing)
- Remove autotune scripts, perf profiles, and prefill schedule planner
- Simplify RuntimePolicy: remove per-arch env overrides, autotune routing

Gemma4:
- Per-layer KV strides: SWA and global attention layers sized independently
- Gemma4 test fixtures updated

Qwen3.5:
- MoE-aware GPU decode path with cached MoE layer keys and scratch buffers
- Refactored decode into encode_qwen35_native_decode_layers with dispatch plan
- Decode cleanup: ~500 lines removed via layer-range abstraction

Metal:
- BF16 batch matmul kernel (bfloat A × f32 B, simdgroup_bfloat8x8, f32 accumulators)
- Q5_0 and Q5_1 batch dequant + matvec shaders
- Pipeline state caching (set_pipeline_cached) across all elementwise dispatches
- Attention kernel cleanup and GDN dispatch improvements

Core:
- Q5_0 / Q5_1 quant module support
- Configurable rayon thread pool (init_global_threadpool_with_count)
- Metrics counter improvements
- Speculative decoding refinements

Docs & packaging:
- License updated to Apache-2.0
- New docs: ENV_VARS, BEST-PRACTICES, PRODUCT-BOUNDARY, LLAMA_SERVER_COMPATIBILITY
- JS SDK description and license updated

+++ crates/ax-engine-core/src/model/arch/qwen3_5/recurrent.rs
@@ -937,7 +937,7 @@ impl Qwen3_5Forward {
                                         encoder,
                                         &bs.norm_buf,
                                         &bs.matmul_in_f16,
-                                        nt * dim as u32,
+                                        (nt as usize * dim) as u32,
                                     );
                                 }
                                 Self::encode_qwen35_batch_projection(
